requirements.txt
sentence-transformers
faiss-cpu
numpy
pandas
requests
scikit-learn
python-dotenv